move CGO out of the client to compile it to many targets
