+++ cli/client/winsize/common.go
@@ -1,0 +1,7 @@
+package winsize
+type WindowSize struct {
+	NRows       uint16
+	NCols       uint16
+	PixelWidth  uint16
+	PixelHeight uint16
+}
