+++ cli/server/main.go
@@ -114,7 +114,7 @@ type Size interface {
 	Size() int64
 }
 
-func setupEnv(user *util.User, runningCommand *runningCommand, authAgentSocketPath string) {
+func setupEnv(user *linux_util.User, runningCommand *runningCommand, authAgentSocketPath string) {
 	// TODO: set the environment like in do_setup_env of https://github.com/openssh/openssh-portable/blob/master/session.c
 	runningCommand.Cmd.Env = append(runningCommand.Cmd.Env,
 		fmt.Sprintf("HOME=%s", user.Dir),
@@ -249,7 +249,7 @@ func forwardTCPInBackground(ctx context.Context, channel ssh3.Channel, conn *net
 	}()
 }
 
-func execCmdInBackground(channel ssh3.Channel, openPty *openPty, user *util.User, runningCommand *runningCommand, authAgentSocketPath string) error {
+func execCmdInBackground(channel ssh3.Channel, openPty *openPty, user *linux_util.User, runningCommand *runningCommand, authAgentSocketPath string) error {
 	setupEnv(user, runningCommand, authAgentSocketPath)
 	if openPty != nil {
 		err := linux_util.StartWithSizeAndPty(&runningCommand.Cmd, openPty.winSize, openPty.pty, openPty.tty)
@@ -377,7 +377,7 @@ func execCmdInBackground(channel ssh3.Channel, openPty *openPty, user *util.User
 	return nil
 }
 
-func newPtyReq(user *util.User, channel ssh3.Channel, request ssh3Messages.PtyRequest, wantReply bool) error {
+func newPtyReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.PtyRequest, wantReply bool) error {
 	var session *runningSession
 	session, ok := runningSessions[channel]
 	if !ok {
@@ -409,11 +409,11 @@ func newPtyReq(user *util.User, channel ssh3.Channel, request ssh3Messages.PtyRe
 	return nil
 }
 
-func newX11Req(user *util.User, channel ssh3.Channel, request ssh3Messages.X11Request, wantReply bool) error {
+func newX11Req(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.X11Request, wantReply bool) error {
 	return fmt.Errorf("%T not implemented", request)
 }
 
-func newCommand(user *util.User, channel ssh3.Channel, command string, args ...string) error {
+func newCommand(user *linux_util.User, channel ssh3.Channel, command string, args ...string) error {
 	var session *runningSession
 	session, ok := runningSessions[channel]
 	if !ok {
@@ -477,24 +477,24 @@ func newCommand(user *util.User, channel ssh3.Channel, command string, args ...s
 	return execCmdInBackground(channel, session.pty, user, session.runningCmd, session.authAgentSocketPath)
 }
 
-func newShellReq(user *util.User, channel ssh3.Channel, wantReply bool) error {
+func newShellReq(user *linux_util.User, channel ssh3.Channel, wantReply bool) error {
 	return newCommand(user, channel, user.Shell)
 }
 
 // similar behaviour to OpenSSH; exec requests are just pasted in the user's shell
-func newCommandInShellReq(user *util.User, channel ssh3.Channel, wantReply bool, command string) error {
+func newCommandInShellReq(user *linux_util.User, channel ssh3.Channel, wantReply bool, command string) error {
 	return newCommand(user, channel, user.Shell, "-c", command)
 }
 
-func newSubsystemReq(user *util.User, channel ssh3.Channel, request ssh3Messages.SubsystemRequest, wantReply bool) error {
+func newSubsystemReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.SubsystemRequest, wantReply bool) error {
 	return fmt.Errorf("%T not implemented", request)
 }
 
-func newWindowChangeReq(user *util.User, channel ssh3.Channel, request ssh3Messages.WindowChangeRequest, wantReply bool) error {
+func newWindowChangeReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.WindowChangeRequest, wantReply bool) error {
 	return fmt.Errorf("%T not implemented", request)
 }
 
-func newSignalReq(user *util.User, channel ssh3.Channel, request ssh3Messages.SignalRequest, wantReply bool) error {
+func newSignalReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.SignalRequest, wantReply bool) error {
 	runningSession, ok := runningSessions[channel]
 	if !ok {
 		return fmt.Errorf("could not find running session for channel %d (conv %d)", channel.ChannelID(), channel.ConversationID())
@@ -520,15 +520,15 @@ func newSignalReq(user *util.User, channel ssh3.Channel, request ssh3Messages.Si
 	return nil
 }
 
-func newExitStatusReq(user *util.User, channel ssh3.Channel, request ssh3Messages.ExitStatusRequest, wantReply bool) error {
+func newExitStatusReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.ExitStatusRequest, wantReply bool) error {
 	return fmt.Errorf("%T not implemented", request)
 }
 
-func newExitSignalReq(user *util.User, channel ssh3.Channel, request ssh3Messages.ExitSignalRequest, wantReply bool) error {
+func newExitSignalReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.ExitSignalRequest, wantReply bool) error {
 	return fmt.Errorf("%T not implemented", request)
 }
 
-func handleUDPForwardingChannel(ctx context.Context, user *util.User, conv *ssh3.Conversation, channel *ssh3.UDPForwardingChannelImpl) error {
+func handleUDPForwardingChannel(ctx context.Context, user *linux_util.User, conv *ssh3.Conversation, channel *ssh3.UDPForwardingChannelImpl) error {
 	// TODO: currently, the rights for socket creation are not checked. The socket is opened with the process's uid and gid
 	// Not sure how to handled that in go since we cannot temporarily change the uid/gid without potentially impacting every
 	// other goroutine
@@ -540,7 +540,7 @@ func handleUDPForwardingChannel(ctx context.Context, user *util.User, conv *ssh3
 	return nil
 }
 
-func handleTCPForwardingChannel(ctx context.Context, user *util.User, conv *ssh3.Conversation, channel *ssh3.TCPForwardingChannelImpl) error {
+func handleTCPForwardingChannel(ctx context.Context, user *linux_util.User, conv *ssh3.Conversation, channel *ssh3.TCPForwardingChannelImpl) error {
 	// TODO: currently, the rights for socket creation are not checked. The socket is opened with the process's uid and gid
 	// Not sure how to handled that in go since we cannot temporarily change the uid/gid without potentially impacting every
 	// other goroutine
@@ -552,7 +552,7 @@ func handleTCPForwardingChannel(ctx context.Context, user *util.User, conv *ssh3
 	return nil
 }
 
-func newDataReq(user *util.User, channel ssh3.Channel, request ssh3Messages.DataOrExtendedDataMessage) error {
+func newDataReq(user *linux_util.User, channel ssh3.Channel, request ssh3Messages.DataOrExtendedDataMessage) error {
 	runningSession, ok := runningSessions[channel]
 	if !ok {
 		return fmt.Errorf("could not find running session for channel %d (conv %d)", channel.ChannelID(), channel.ConversationID())
@@ -637,7 +637,7 @@ func listenAndAcceptAuthSockets(cancel context.CancelCauseFunc, conversation *ss
 	}
 }
 
-func openAgentSocketAndForwardAgent(parent context.Context, conv *ssh3.Conversation, user *util.User) (string, error) {
+func openAgentSocketAndForwardAgent(parent context.Context, conv *ssh3.Conversation, user *linux_util.User) (string, error) {
 	ctx, cancel := context.WithCancelCause(parent)
 	sockPath, err := linux_util.NewUnixSocketPath()
 	if err != nil {
@@ -738,7 +738,7 @@ func main() {
 
 		mux := http.NewServeMux()
 		ssh3Server := ssh3.NewServer(30000, 10, &server, func(authenticatedUsername string, conv *ssh3.Conversation) error {
-			authenticatedUser, err := util.GetUser(authenticatedUsername)
+			authenticatedUser, err := linux_util.GetUser(authenticatedUsername)
 			if err != nil {
 				return err
 			}
